+++ parsers/json.js
@@ -1,1 +1,0 @@
-export const json = (...args) => JSON.parse(String.raw.apply(null, args).trim());