+++ parsers/text.js
@@ -1,1 +1,0 @@
-export const text = (...args) => String.raw.apply(null, args).trim();