+++ rollup.config.js
@@ -1,35 +1,30 @@
 import terser from '@rollup/plugin-terser';
 
-export default [{
-	input: 'sanitizer.js',
-	output: [{
-		file: 'sanitizer.cjs',
-		format: 'cjs',
-	}, {
-		file: 'sanitizer.mjs',
-		format: 'module',
-	}],
-}, {
-	input: 'polyfill.js',
-	output: [{
-		file: 'polyfill.cjs',
-		format: 'cjs',
-	}, {
-		file: 'polyfill.mjs',
-		format: 'module',
-	}, {
-		file: 'polyfill.min.js',
-		format: 'iife',
-		sourcemap: true,
-		plugins: [terser()],
-	}],
-}, {
-	input: 'parsers.js',
-	output: [{
-		file: 'parsers.cjs',
-		format: 'cjs',
-	}, {
-		file: 'parsers.mjs',
-		format: 'module',
-	}],
-}];
+const modules = [
+	'sanitizer', 'config', 'config/base', 'config/complete', 'config/html',
+	'config/html', 'config/html', 'config/mathml', 'config/svg', 'config/global',
+	'namespaces', 'config-utils',
+];
+
+export default [
+	{
+		input: 'polyfill.js',
+		output: [{
+			file: 'polyfill.cjs',
+			format: 'cjs',
+		}, {
+			file: 'polyfill.min.js',
+			format: 'iife',
+			sourcemap: true,
+			plugins: [terser()],
+		}],
+	},
+	...modules.map(path => ({
+		input: `${path}.js`,
+		external: () => true,
+		output: {
+			file: `${path}.cjs`,
+			format: 'cjs',
+		},
+	}))
+];