+++ sanitizer.js
@@ -2,4 +2,3 @@ export { htmlConfig, svgConfig, mathMLConfig, baseConfig, completeConfig } from 
 export { normalizeConfig } from './config-utils.js';
 export { setHTML, parseHTML, sanitize } from './sanitize.js';
 export { HTML as HTMLNS, SVG as SVGNS, MATHML as MATHMLNS } from './namespaces.js';
-export { html, css, svg, math, xml, json, text, createHTMLParser, createCSSParser } from './parsers.js';