+++ rollup.config.js
@@ -1,14 +1,21 @@
 import terser from '@rollup/plugin-terser';
+import nodeResolve from '@rollup/plugin-node-resolve';
+
+const t = terser();
+const nr = nodeResolve();
+
+const configs = ['base', 'html', 'svg', 'mathml', 'complete'];
 
 const modules = [
 	'sanitizer', 'config', 'config/base', 'config/complete', 'config/html',
 	'config/html', 'config/html', 'config/mathml', 'config/svg', 'config/global',
-	'namespaces', 'config-utils',
+	'namespaces', 'config-utils', 'sanitize',
 ];
 
 export default [
 	{
 		input: 'polyfill.js',
+		plugins: [nr],
 		output: [{
 			file: 'polyfill.cjs',
 			format: 'cjs',
@@ -16,7 +23,7 @@ export default [
 			file: 'polyfill.min.js',
 			format: 'iife',
 			sourcemap: true,
-			plugins: [terser()],
+			plugins: [t],
 		}],
 	},
 	...modules.map(path => ({
@@ -26,5 +33,15 @@ export default [
 			file: `${path}.cjs`,
 			format: 'cjs',
 		},
-	}))
+	})),
+	...configs.map(path => ({
+		input: `config/${path}.js`,
+		plugins: [nr],
+		output: {
+			file: `config/${path}.min.js`,
+			format: 'module',
+			sourcemap: true,
+			plugins: [t],
+		}
+	})),
 ];